+++ crates/matrix-sdk/src/sliding_sync/builder.rs
@@ -40,6 +40,8 @@ pub struct SlidingSyncBuilder {
     rooms: BTreeMap<OwnedRoomId, SlidingSyncRoom>,
     poll_timeout: Duration,
     network_timeout: Duration,
+    #[cfg(feature = "e2e-encryption")]
+    share_pos: bool,
 }
 
 impl SlidingSyncBuilder {
@@ -62,6 +64,8 @@ impl SlidingSyncBuilder {
                 rooms: BTreeMap::new(),
                 poll_timeout: Duration::from_secs(30),
                 network_timeout: Duration::from_secs(30),
+                #[cfg(feature = "e2e-encryption")]
+                share_pos: false,
             })
         }
     }
@@ -222,14 +226,26 @@ impl SlidingSyncBuilder {
         self
     }
 
+    /// Should the sliding sync instance share its sync position through
+    /// storage?
+    ///
+    /// In general, sliding sync instances will cache the sync position (`pos`
+    /// field in the request) in internal memory. It can be useful, in
+    /// multi-process scenarios, to save it into some shared storage so that one
+    /// sliding sync instance running across two different processes can
+    /// continue with the same sync position it had before being stopped.
+    #[cfg(feature = "e2e-encryption")]
+    pub fn share_pos(mut self) -> Self {
+        self.share_pos = true;
+        self
+    }
+
     /// Build the Sliding Sync.
     ///
     /// If `self.storage_key` is `Some(_)`, load the cached data from cold
     /// storage.
     pub async fn build(self) -> Result<SlidingSync> {
         let client = self.client;
-
-        let mut delta_token = None;
 
         let (internal_channel_sender, _internal_channel_receiver) = channel(8);
 
@@ -242,14 +258,24 @@ impl SlidingSyncBuilder {
         }
 
         // Reload existing state from the cache.
-        restore_sliding_sync_state(
-            &client,
-            &self.storage_key,
-            &lists,
-            &mut delta_token,
-            &mut None, // to_device_token is ignored here
-        )
-        .await?;
+        let restored_fields =
+            restore_sliding_sync_state(&client, &self.storage_key, &lists).await?;
+
+        let (delta_token, pos) = if let Some(fields) = restored_fields {
+            #[cfg(feature = "e2e-encryption")]
+            let pos = if self.share_pos { fields.pos } else { None };
+            #[cfg(not(feature = "e2e-encryption"))]
+            let pos = None;
+
+            (fields.delta_token, pos)
+        } else {
+            (None, None)
+        };
+
+        #[cfg(feature = "e2e-encryption")]
+        let share_pos = self.share_pos;
+        #[cfg(not(feature = "e2e-encryption"))]
+        let share_pos = false;
 
         let rooms = AsyncRwLock::new(self.rooms);
         let lists = AsyncRwLock::new(lists);
@@ -264,14 +290,12 @@ impl SlidingSyncBuilder {
 
             client,
             storage_key: self.storage_key,
+            share_pos,
 
             lists,
             rooms,
 
-            position: Arc::new(AsyncMutex::new(SlidingSyncPositionMarkers {
-                pos: None,
-                delta_token,
-            })),
+            position: Arc::new(AsyncMutex::new(SlidingSyncPositionMarkers { pos, delta_token })),
             past_positions: StdRwLock::new(RingBuffer::new(20)),
 
             sticky: StdRwLock::new(SlidingSyncStickyManager::new(