+++ crates/matrix-sdk/src/sliding_sync/cache.rs
@@ -15,7 +15,10 @@ use super::{
     FrozenSlidingSync, FrozenSlidingSyncList, SlidingSync, SlidingSyncList,
     SlidingSyncPositionMarkers,
 };
-use crate::{sliding_sync::SlidingSyncListCachePolicy, Client, Result};
+use crate::{
+    sliding_sync::{FrozenSlidingSyncPos, SlidingSyncListCachePolicy},
+    Client, Result,
+};
 
 /// Be careful: as this is used as a storage key; changing it requires migrating
 /// data!
@@ -57,9 +60,17 @@ async fn clean_storage(
     for list_name in lists.keys() {
         invalidate_cached_list(storage, storage_key, list_name).await;
     }
-    let _ = storage
-        .remove_custom_value(format_storage_key_for_sliding_sync(storage_key).as_bytes())
-        .await;
+    let instance_storage_key = format_storage_key_for_sliding_sync(storage_key);
+    let _ = storage.remove_custom_value(instance_storage_key.as_bytes()).await;
+
+    #[cfg(feature = "e2e-encryption")]
+    if let Some(olm_machine) = &*client.olm_machine().await {
+        // Invalidate the value stored for the TERRIBLE HACK.
+        let _ = olm_machine
+            .store()
+            .set_custom_value(&instance_storage_key, "".as_bytes().to_vec())
+            .await;
+    }
 }
 
 /// Store the `SlidingSync`'s state in the storage.
@@ -81,6 +92,18 @@ pub(super) async fn store_sliding_sync_state(
             serde_json::to_vec(&FrozenSlidingSync::new(position).await)?,
         )
         .await?;
+
+    #[cfg(feature = "e2e-encryption")]
+    {
+        // FIXME (TERRIBLE HACK): we want to save `pos` in a cross-process safe manner,
+        // with both processes sharing the same database backend; that needs to
+        // go in the crypto process store at the moment, but should be fixed
+        // later on.
+        if let Some(olm_machine) = &*sliding_sync.inner.client.olm_machine().await {
+            let pos_blob = serde_json::to_vec(&FrozenSlidingSyncPos { pos: position.pos.clone() })?;
+            olm_machine.store().set_custom_value(&instance_storage_key, pos_blob).await?;
+        }
+    }
 
     // Write every `SlidingSyncList` that's configured for caching into the store.
     let frozen_lists = {
@@ -158,6 +181,14 @@ pub(super) async fn restore_sliding_sync_list(
     Ok(None)
 }
 
+/// Fields restored during `restore_sliding_sync_state`.
+#[derive(Default)]
+pub(super) struct RestoredFields {
+    pub delta_token: Option<String>,
+    pub to_device_token: Option<String>,
+    pub pos: Option<String>,
+}
+
 /// Restore the `SlidingSync`'s state from what is stored in the storage.
 ///
 /// If one cache is obsolete (corrupted, and cannot be deserialized or
@@ -166,26 +197,27 @@ pub(super) async fn restore_sliding_sync_state(
     client: &Client,
     storage_key: &str,
     lists: &BTreeMap<String, SlidingSyncList>,
-    delta_token: &mut Option<String>,
-    to_device_token: &mut Option<String>,
-) -> Result<()> {
+) -> Result<Option<RestoredFields>> {
     let _timer = timer!(format!("loading sliding sync {storage_key} state from DB"));
+
+    let mut restored_fields = RestoredFields::default();
 
     #[cfg(feature = "e2e-encryption")]
     if let Some(olm_machine) = &*client.olm_machine().await {
         match olm_machine.store().next_batch_token().await? {
             Some(token) => {
-                *to_device_token = Some(token);
+                restored_fields.to_device_token = Some(token);
             }
             None => trace!("No `SlidingSync` in the crypto-store cache"),
         }
     }
 
     let storage = client.store();
+    let instance_storage_key = format_storage_key_for_sliding_sync(storage_key);
 
     // Preload the `SlidingSync` object from the cache.
     match storage
-        .get_custom_value(format_storage_key_for_sliding_sync(storage_key).as_bytes())
+        .get_custom_value(instance_storage_key.as_bytes())
         .await?
         .map(|custom_value| serde_json::from_slice::<FrozenSlidingSync>(&custom_value))
     {
@@ -194,10 +226,27 @@ pub(super) async fn restore_sliding_sync_state(
             trace!("Successfully read the `SlidingSync` from the cache");
             // Only update the to-device token if we failed to read it from the crypto store
             // above.
-            if to_device_token.is_none() {
-                *to_device_token = to_device_since;
+            if restored_fields.to_device_token.is_none() {
+                restored_fields.to_device_token = to_device_since;
             }
-            *delta_token = frozen_delta_token;
+
+            restored_fields.delta_token = frozen_delta_token;
+
+            #[cfg(feature = "e2e-encryption")]
+            {
+                if let Some(olm_machine) = &*client.olm_machine().await {
+                    if let Ok(Some(blob)) =
+                        olm_machine.store().get_custom_value(&instance_storage_key).await
+                    {
+                        if let Ok(frozen_pos) =
+                            serde_json::from_slice::<FrozenSlidingSyncPos>(&blob)
+                        {
+                            trace!("Successfully read the `Sliding Sync` pos from the crypto store cache");
+                            restored_fields.pos = frozen_pos.pos;
+                        }
+                    }
+                }
+            }
         }
 
         // `SlidingSync` has been found, but it wasn't possible to deserialize it. It's
@@ -214,7 +263,7 @@ pub(super) async fn restore_sliding_sync_state(
             // Let's clear everything and stop here.
             clean_storage(client, storage_key, lists).await;
 
-            return Ok(());
+            return Ok(None);
         }
 
         None => {
@@ -222,7 +271,7 @@ pub(super) async fn restore_sliding_sync_state(
         }
     }
 
-    Ok(())
+    Ok(Some(restored_fields))
 }
 
 #[cfg(test)]
@@ -413,9 +462,11 @@ mod tests {
 
         // Emulate some data to be cached.
         let delta_token = "delta_token".to_owned();
+        let pos = "pos".to_owned();
         {
             let mut position_guard = sliding_sync.inner.position.lock().await;
             position_guard.delta_token = Some(delta_token.clone());
+            position_guard.pos = Some(pos.clone());
 
             // Then, we can correctly cache the sliding sync instance.
             store_sliding_sync_state(&sliding_sync, &position_guard).await?;
@@ -436,19 +487,13 @@ mod tests {
         // Ok, forget about the sliding sync, let's recreate one from scratch.
         drop(sliding_sync);
 
-        let mut read_delta_token = None;
-
-        restore_sliding_sync_state(
-            &client,
-            &storage_key_prefix,
-            &[].into(),
-            &mut read_delta_token,
-            &mut None,
-        )
-        .await?;
+        let restored_fields = restore_sliding_sync_state(&client, &storage_key_prefix, &[].into())
+            .await?
+            .expect("must have restored sliding sync fields");
 
         // After restoring, the delta token and to-device token could be read.
-        assert_eq!(read_delta_token.unwrap(), delta_token);
+        assert_eq!(restored_fields.delta_token.unwrap(), delta_token);
+        assert_eq!(restored_fields.pos.unwrap(), pos);
 
         // Test the "migration" path: assume a missing to-device token in crypto store,
         // but present in a former state store.
@@ -474,22 +519,15 @@ mod tests {
             )
             .await?;
 
-        let mut read_delta_token = None;
-        let mut read_to_device_token = None;
+        let restored_fields = restore_sliding_sync_state(&client, &storage_key_prefix, &[].into())
+            .await?
+            .expect("must have restored fields");
 
-        restore_sliding_sync_state(
-            &client,
-            &storage_key_prefix,
-            &[].into(),
-            &mut read_delta_token,
-            &mut read_to_device_token,
-        )
-        .await?;
-
-        // After restoring, the delta token and to-device token could be read, even if
-        // they came both from the state store.
-        assert_eq!(read_delta_token.unwrap(), delta_token);
-        assert_eq!(read_to_device_token.unwrap(), to_device_token);
+        // After restoring, the delta token, the to-device since token, and stream
+        // position could be read from the state store.
+        assert_eq!(restored_fields.delta_token.unwrap(), delta_token);
+        assert_eq!(restored_fields.to_device_token.unwrap(), to_device_token);
+        assert_eq!(restored_fields.pos.unwrap(), pos);
 
         Ok(())
     }